changed price from double to decimal

+++ FakeNFT/Models/Network/Nft.swift
@@ -4,6 +4,6 @@ struct Nft: Decodable {
     let id: String
     let name: String
     let rating: Int
-    let price: Double
+    let price: Decimal
     let images: [URL]
 }
